+++ samples/projectile-basic/main.cpp
@@ -101,7 +101,7 @@ void setupProjectileDisplay(ImGuiSystem& imgui, ecs::World& world, BulletPhysics
             {
                 if (force && force->isActive())
                 {
-                    float magnitude = force->getForce().length();
+                    double magnitude = force->getForce().length();
                     ImGui::Text("   %s: %.8f N", force->getSymbol().c_str(), magnitude);
                 }
             }
@@ -170,9 +170,9 @@ int main()
 
     // configure physics world
     physicsWorld.addForce(std::make_unique<BulletPhysics::dynamics::forces::Gravity>());
-    physicsWorld.addEnvironment(std::make_unique<BulletPhysics::dynamics::environment::Atmosphere>(280.0f, 100000.0f));                      // t_0 = 280 K, p_0 = 100.000 Pa
+    physicsWorld.addEnvironment(std::make_unique<BulletPhysics::dynamics::environment::Atmosphere>(280.0, 100000.0));                      // t_0 = 280 K, p_0 = 100.000 Pa
     physicsWorld.addEnvironment(std::make_unique<BulletPhysics::dynamics::environment::Humidity>(60));                                       // relative humidity = 60%
-    physicsWorld.addEnvironment(std::make_unique<BulletPhysics::dynamics::environment::Wind>(BulletPhysics::math::Vec3{0.0f, 0.0f, 2.0f}));   // wind velocity = 2 m/s
+    physicsWorld.addEnvironment(std::make_unique<BulletPhysics::dynamics::environment::Wind>(BulletPhysics::math::Vec3{0.0, 0.0, 2.0}));   // wind velocity = 2 m/s
     physicsWorld.addEnvironment(std::make_unique<BulletPhysics::dynamics::environment::Geographic>(BulletPhysics::math::deg2rad(48.1482), BulletPhysics::math::deg2rad(17.1067))); // Bratislava coordinates
     physicsWorld.addForce(std::make_unique<BulletPhysics::dynamics::forces::drag::Drag>());
     physicsWorld.addForce(std::make_unique<BulletPhysics::dynamics::forces::Coriolis>());
